Add video processing capabilities to ContentManagerView
Implemented a file importer and video processing functionality in ContentManagerView, utilizing CoreML and video segmentation techniques. This update enables scene detection and analysis of video files, enhancing content management with detailed frame processing and match detection. The logic previously residing in ContentView has been removed to streamline code structure.
